Added support for more detailed tcp/ip error messages.

+++ pest++/yamr/Transformable.cpp
@@ -57,7 +57,7 @@ Transformable::Transformable(const vector<string> &names, const Eigen::VectorXd 
         throw PestIndexError("Transformable::Transformable(const vector<string> &names, Eigen::VectorXd &values)", 
 			"size of names vector does not match the size of the values vector");
 	}
-    size_t len = min(names.size(), values.size());
+    size_t len = min(size_t(names.size()), size_t(values.size()));
     for (size_t i=0; i<len; ++i)
 	{
         items[names[i]] = values(i);
